requirements.txt
bcrypt==3.2.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-pubsub==2.2.0
google-cloud-secret-manager==2.1.0
google-cloud-storage==1.35.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.34.0
idna==2.10
libcst==0.3.16
mypy-extensions==0.4.3
paramiko==2.7.2
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pysftp==0.2.9
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
